requirements.txt
Django==4.2.16
django-environ==0.11.2
whitenoise==6.7.0
Pillow==10.4.0
python-dateutil==2.9.0.post0
pypdf==4.3.1
pytesseract==0.3.10
pdf2image==1.17.0
